Removing shared object.
